requirements.txt
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
WooCommerce==3.0.0
